backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
anthropic[bedrock]==0.28.0
Pillow==10.3.0
python-dotenv==1.0.1
google-generativeai==0.6.0